+++ include/InteractiveToolkit/MathCore/impl/mat2_all_no_optimization.h
@@ -55,11 +55,11 @@ namespace MathCore
 
     public:
 
-        static constexpr size_t rows = 2;
-		static constexpr size_t cols = 2;
+        static constexpr int rows = 2;
+		static constexpr int cols = 2;
 
-        static constexpr size_t array_count = 4;
-        static constexpr size_t array_stride = 2;
+        static constexpr int array_count = 4;
+        static constexpr int array_stride = 2;
 
         using type = self_type;
 		using element_type = _BaseType;
