+++ include/InteractiveToolkit/MathCore/impl/mat4_all_no_optimization.h
@@ -54,11 +54,11 @@ namespace MathCore
 
 	public:
 
-		static constexpr size_t rows = 4;
-		static constexpr size_t cols = 4;
+		static constexpr int rows = 4;
+		static constexpr int cols = 4;
 
-		static constexpr size_t array_count = 16;
-        static constexpr size_t array_stride = 4;
+		static constexpr int array_count = 16;
+        static constexpr int array_stride = 4;
 
 		using type = self_type;
 		using element_type = _BaseType;
